Tools/ArtPipeline/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0
scipy>=1.11.0
tqdm>=4.65.0
